requirements.txt
langgraph>=0.0.26
langgraph-checkpoint-postgres>=1.0.1
psycopg[binary,pool]>=3.1.18
langchain-openai>=0.1.3
pydantic>=2.6.3
python-telegram-bot[job-queue]>=21.0.1
structlog>=24.1.0
python-dotenv>=1.0.1
langsmith>=0.1.23
google-api-python-client>=2.122.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
tiktoken>=0.7.0
